w3c-reports/12092014/efl/runned-tests/eventsource/request-credentials-expected.txt


FAIL EventSource: credentials: credentials enabled assert_equals: cookie status expected "second COOKIE" but got "second NO_COOKIE"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/eventsource/request-credentials.htm:26:30
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
PASS EventSource: credentials: credentials disabled 
PASS EventSource: credentials: credentials default 
